+++ packages/linzhe-tools/index.ts
@@ -1,1 +1,2 @@
 export * from '@linzhe-tools/components'
+export * from '@linzhe-tools/shared'